fix unfollow;

+++ src/resolvers/Mutation/unfollow.ts
@@ -1,13 +1,26 @@
-import { MutationResolvers } from "~/generated/graphql";
-import { ForbiddenError } from "apollo-server-express";
+import { MutationResolvers } from '~/generated/graphql';
+import { ForbiddenError } from 'apollo-server-express';
 
-export const unfollow: MutationResolvers["unfollow"] = async (
+export const unfollow: MutationResolvers['unfollow'] = async (
   _,
   { followeeId },
   { prisma, requestUser }
 ) => {
   if (!requestUser) {
-    throw new ForbiddenError("auth error");
+    throw new ForbiddenError('auth error');
+  }
+
+  const targetUser = await prisma.user.findUnique({
+    where: {
+      id: followeeId,
+    },
+  });
+
+  if (!targetUser) {
+    return {
+      __typename: 'Deleted',
+      message: 'ユーザーが見つかりません',
+    };
   }
 
   const result = await prisma.follow.delete({
@@ -24,6 +37,7 @@ export const unfollow: MutationResolvers["unfollow"] = async (
 
   return {
     ...result.followee,
+    __typename: 'User',
     follow: false,
   };
 };
